# rlp: minor optimizations for slice/array encoding